requirements.txt
psycopg2==2.4.5
gunicorn==0.14.3
raven==1.9.4
blinker==1.2

flask==0.9
flask-sqlalchemy

static
